gateway/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
websockets
psycopg2
pydantic
pyjwt